add geo mapping

+++ app/src/main/java/dto/UberPickDTO.java
@@ -1,8 +1,10 @@
 package dto;
 
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import dto.geo.UberPickGeoDTO;
 
 import java.time.LocalDateTime;
+import java.util.List;
 
 /**
  * Created by Adrian on 2017-03-21.
@@ -10,21 +12,23 @@ import java.time.LocalDateTime;
 public class UberPickDTO {
     @JsonSerialize(using = StandardDateSerializer.class)
     private final LocalDateTime dateTime;
-    private final double[] location = {0,0};
+    private final float[] location = {0,0};
     private final String base;
 
-    public UberPickDTO(LocalDateTime dateTime, double latitude, double longitude, String base) {
+    public UberPickDTO(LocalDateTime dateTime, float latitude, float longitude, String base) {
         this.dateTime = dateTime;
         this.location[0] = latitude;
         this.location[1] = longitude;
         this.base = base;
     }
 
+
+
     public LocalDateTime getDateTime() {
         return dateTime;
     }
 
-    public double[] getLocation() {
+    public float[] getLocation() {
         return location;
     }
