+++ app/src/main/java/dto/UberPickParser.java
@@ -15,8 +15,8 @@ public class UberPickParser {
     public UberPickDTO parse(String line) {
         String[] split = line.replace("\"","").split(",");
         LocalDateTime dateTime = LocalDateTime.parse(split[0], formatter);
-        double latitude = Double.parseDouble(split[1]);
-        double longitude = Double.parseDouble(split[2]);
+        float latitude = Float.parseFloat(split[1]);
+        float  longitude = Float.parseFloat(split[2]);
         String base = split[3];
         return new UberPickDTO(dateTime, latitude, longitude, base);
     }
